+++ src/main/java/roomescape/domain/reservation/controller/ReservationTimeController.java
@@ -1,17 +1,11 @@
 package roomescape.domain.reservation.controller;
 
-import java.net.URI;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservation.request.ReservationTimeCreateRequest;
 import roomescape.domain.reservation.response.ReservationTimeResponse;
 import roomescape.domain.reservation.response.ReservationTimesResponse;
 import roomescape.domain.reservation.service.ReservationTimeService;
@@ -31,18 +25,5 @@ public class ReservationTimeController {
     public ResponseEntity<ReservationTimesResponse> findAll() {
         List<ReservationTimeResponse> times = reservationTimeService.findAllReservationTimes();
         return ResponseEntity.ok(new ReservationTimesResponse(times));
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponse> save(@RequestBody ReservationTimeCreateRequest request) {
-        ReservationTimeResponse response = reservationTimeService.saveReservationTime(request);
-        return ResponseEntity.created(URI.create("/times/" + response.id()))
-                .body(response);
-    }
-
-    @DeleteMapping("/{reservationTimeId}")
-    public ResponseEntity<Void> deleteById(@PathVariable Long reservationTimeId) {
-        reservationTimeService.deleteReservationTimeBy(reservationTimeId);
-        return ResponseEntity.ok().build();
     }
 }
